Extract worker scaffold test helpers (#66)
Move shared Celery worker scaffold test helpers into `tests/conftest.py`
so `tests/test_worker_service_scaffold.py` stays below the module-line
limit without changing pylint configuration.

Keep the worker scaffold tests behaviourally unchanged by reusing the new
fixture and helper symbols from the existing test module.
